+++ src/main/scala/ai/tripl/arc/extract/JSONExtract.scala
@@ -164,12 +164,11 @@ object JSONExtractStage {
       if (arcContext.isStreaming) {
         stage.input match {
           case Right(glob) => {
-
             optionSchema match {
               case Some(schema) => {
                 stage.watermark match {
-                  case Some(watermark) => Right(spark.readStream.options(options).schema(schema).json(glob).withWatermark(watermark.eventTime, watermark.delayThreshold))
-                  case None => Right(spark.readStream.options(options).schema(schema).json(glob))
+                  case Some(watermark) => Right(spark.readStream.options(options).schema(schema).format("json").load(glob).withWatermark(watermark.eventTime, watermark.delayThreshold))
+                  case None => Right(spark.readStream.options(options).schema(schema).format("json").load(glob))
                 }
               }
               case None => throw new Exception("JSONExtract requires 'schemaURI' or 'schemaView' to be set if Arc is running in streaming mode.")
@@ -190,70 +189,23 @@ object JSONExtractStage {
       } else {
         stage.input match {
           case Right(glob) =>
-            if (stage.basePath.isEmpty) {
-              // spark does not cope well reading many small files into json directly from hadoop file systems
-              // by reading first as text time drops by ~75%
-              // this will not throw an error for empty directory (but will for missing directory)
-              try {
-                if (stage.settings.multiLine) {
-
-                  // if multiLine then remove the crlf delimiter so it is read as a full object per file
-                  val oldDelimiter = spark.sparkContext.hadoopConfiguration.get("textinputformat.record.delimiter")
-                  val newDelimiter = s"${0x0 : Char}"
-                  // temporarily remove the delimiter so all the data is loaded as a single line
-                  spark.sparkContext.hadoopConfiguration.set("textinputformat.record.delimiter", newDelimiter)
-
-                  // read the file but do not cache. caching will break the input_file_name() function
-                  val textFile = spark.sparkContext.textFile(glob)
-
-                  val json = optionSchema match {
-                    case Some(schema) => Right(spark.read.options(options).schema(schema).json(textFile.toDS))
-                    case None => Right(spark.read.options(options).json(textFile.toDS))
-                  }
-
-                  // reset delimiter
-                  if (oldDelimiter == null) {
-                    spark.sparkContext.hadoopConfiguration.unset("textinputformat.record.delimiter")
-                  } else {
-                    spark.sparkContext.hadoopConfiguration.set("textinputformat.record.delimiter", oldDelimiter)
-                  }
-
-                  json
-                } else {
-                  // read the file but do not cache. caching will break the input_file_name() function
-                  val textFile = spark.sparkContext.textFile(glob)
-
-                  val json = optionSchema match {
-                    case Some(schema) => Right(spark.read.options(options).schema(schema).json(textFile.toDS))
-                    case None => Right(spark.read.options(options).json(textFile.toDS))
-                  }
-
-                  json
-                }
-              } catch {
-                case e: org.apache.hadoop.mapred.InvalidInputException if (e.getMessage.contains("matches 0 files")) =>
-                  Left(FileNotFoundExtractError(Option(glob)))
-                case e: Exception => throw e
+            try {
+              optionSchema match {
+                case Some(schema) => Right(spark.read.options(options).schema(schema).format("json").load(glob))
+                case None => Right(spark.read.options(options).format("json").load(glob))
               }
-            } else {
-              // if basePath is defined the textFile hack cannot be used
-              try {
-                optionSchema match {
-                  case Some(schema) => Right(spark.read.options(options + ("multiLine" -> stage.settings.multiLine.toString)).schema(schema).json(glob))
-                  case None => Right(spark.read.options(options + ("multiLine" -> stage.settings.multiLine.toString)).json(glob))
-                }
-              } catch {
-                case e: AnalysisException if (e.getMessage == "Unable to infer schema for JSON. It must be specified manually.;") =>
-                  Left(FileNotFoundExtractError(Option(glob)))
-                case e: AnalysisException if (e.getMessage.contains("Path does not exist")) =>
-                  Left(PathNotExistsExtractError(Option(glob)))
-                case e: Exception => throw e
-              }
+            } catch {
+              case e: AnalysisException if (e.getMessage == "Unable to infer schema for JSON. It must be specified manually.;") =>
+                Left(FileNotFoundExtractError(Option(glob)))
+              case e: AnalysisException if (e.getMessage.contains("Path does not exist")) =>
+                Left(PathNotExistsExtractError(Option(glob)))
+              case e: Exception => throw e
             }
           case Left(view) => {
+            val inputView = spark.table(view)
             stage.inputField match {
-              case Some(inputField) => Right(spark.read.options(options).json(spark.table(view).select(col(inputField).as("value")).as[String]))
-              case None => Right(spark.read.options(options).json(spark.table(view).as[String]))
+              case Some(inputField) => Right(spark.read.options(options).json(inputView.select(col(inputField).as("value")).as[String]))
+              case None => Right(spark.read.options(options).json(inputView.as[String]))
             }
           }
         }
@@ -263,7 +215,6 @@ object JSONExtractStage {
         override val detail = stage.stageDetail
       }
     }
-
 
     // if incoming dataset has 0 columns then try to create empty dataset with correct schema
     // or throw enriched error message