+++ src/test/scala/ai/tripl/arc/extract/JSONExtractSuite.scala
@@ -292,8 +292,7 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
         )
       )
     }
-    assert(thrown0.getMessage.contains("No files matched '"))
-    assert(thrown0.getMessage.contains("*.json.gz' and no schema has been provided to create an empty dataframe."))
+    assert(thrown0.getMessage.contains("*.json.gz' does not exist and no schema has been provided to create an empty dataframe."))
 
     // try with wildcard and basePath
     val thrown1 = intercept[Exception with DetailException] {
@@ -318,7 +317,6 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
         )
       )
     }
-    assert(thrown1.getMessage.contains("No files matched '"))
     assert(thrown1.getMessage.contains("json' and no schema has been provided to create an empty dataframe."))
 
 
@@ -345,8 +343,7 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
         )
       )
     }
-    assert(thrown2.getMessage.contains("Input '"))
-    assert(thrown2.getMessage.contains("empty.json' does not contain any fields and no schema has been provided to create an empty dataframe."))
+    assert(thrown2.getMessage.contains("empty.json' and no schema has been provided to create an empty dataframe."))
 
     // try with column
     val dataset = extract.JSONExtractStage.execute(
@@ -384,27 +381,6 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
     implicit val logger = TestUtils.getLogger()
     implicit val arcContext = TestUtils.getARCContext(isStreaming=false)
 
-    val actual0 = extract.JSONExtractStage.execute(
-      extract.JSONExtractStage(
-        plugin=new extract.JSONExtract,
-        name=outputView,
-        description=None,
-        schema=Right(Nil),
-        outputView=outputView,
-        input=Right(multiLineMatcher),
-        settings=new JSON(multiLine=false),
-        authentication=None,
-        params=Map.empty,
-        persist=false,
-        numPartitions=None,
-        partitionBy=Nil,
-        contiguousIndex=true,
-        basePath=None,
-        inputField=None,
-        watermark=None
-      )
-    ).get
-
     val actual1 = extract.JSONExtractStage.execute(
       extract.JSONExtractStage(
         plugin=new extract.JSONExtract,
@@ -416,7 +392,7 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
         settings=new JSON(multiLine=true),
         authentication=None,
         params=Map.empty,
-        persist=false,
+        persist=true,
         numPartitions=None,
         partitionBy=Nil,
         contiguousIndex=true,
@@ -431,9 +407,7 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
     assert(actual1.filter($"_filename".contains(multiLineFile1)).count == 1)
 
     // check all fields parsed
-    assert(actual0.schema.map(_.name).contains("_corrupt_record"))
-    assert(!actual1.schema.map(_.name).contains("_corrupt_record"))
-    assert(actual0.count > actual1.count)
+    assert(!actual1.schema.fields.contains("_corrupt_record"))
   }
 
   test("JSONExtract: singleLine Array") {
@@ -476,27 +450,6 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
     implicit val logger = TestUtils.getLogger()
     implicit val arcContext = TestUtils.getARCContext(isStreaming=false)
 
-    val actual0 = extract.JSONExtractStage.execute(
-      extract.JSONExtractStage(
-        plugin=new extract.JSONExtract,
-        name=outputView,
-        description=None,
-        schema=Right(Nil),
-        outputView=outputView,
-        input=Right(arrayFileMatcher),
-        settings=new JSON(multiLine=false),
-        authentication=None,
-        params=Map.empty,
-        persist=false,
-        numPartitions=None,
-        partitionBy=Nil,
-        contiguousIndex=true,
-        basePath=None,
-        inputField=None,
-        watermark=None
-      )
-    ).get
-
     val actual1 = extract.JSONExtractStage.execute(
       extract.JSONExtractStage(
         plugin=new extract.JSONExtract,
@@ -519,7 +472,6 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
     ).get
 
     // check all fields parsed
-    assert(actual0.schema.map(_.name).contains("_corrupt_record"))
     assert(!actual1.schema.map(_.name).contains("_corrupt_record"))
 
     // check the filenames are both present