+++ src/test/scala/ai/tripl/arc/extract/TextExtractSuite.scala
@@ -10,6 +10,7 @@ import collection.JavaConverters._
 import org.apache.commons.io.FileUtils
 import org.apache.commons.io.IOUtils
 import org.apache.spark.sql._
+import org.apache.spark.sql.expressions.Window
 import org.apache.spark.sql.functions._
 
 import ai.tripl.arc.api._
@@ -79,7 +80,7 @@ class TextExtractSuite extends FunSuite with BeforeAndAfter {
         assert(df.count == 1)
       }
     }
-  }  
+  }
 
   test("TextExtract: end-to-end inputView") {
     implicit val spark = session
@@ -116,7 +117,7 @@ class TextExtractSuite extends FunSuite with BeforeAndAfter {
         assert(df.count == 1)
       }
     }
-  }  
+  }
 
   test("TextExtract: multiLine false") {
     implicit val spark = session
@@ -205,6 +206,42 @@ class TextExtractSuite extends FunSuite with BeforeAndAfter {
     }
     assert(thrown0.getMessage.contains("Path '"))
     assert(thrown0.getMessage.contains("empty.text' does not exist and no schema has been provided to create an empty dataframe."))
+  }
+
+  test("TextExtract: _index") {
+    implicit val spark = session
+    import spark.implicits._
+    implicit val logger = TestUtils.getLogger()
+    implicit val arcContext = TestUtils.getARCContext(isStreaming=false)
+
+    val actual = extract.TextExtractStage.execute(
+      extract.TextExtractStage(
+        plugin=new extract.TextExtract,
+        name="dataset",
+        description=None,
+        schema=Right(List.empty),
+        outputView=outputView,
+        input=Right(targetDirectory),
+        authentication=None,
+        persist=false,
+        numPartitions=None,
+        contiguousIndex=true,
+        multiLine=false,
+        basePath=None,
+        params=Map.empty,
+        watermark=None
+      )
+    ).get
+
+    val window = Window.partitionBy("_filename").orderBy("_monotonically_increasing_id")
+    val expected = spark.read.format("text").load(targetDirectory)
+      .withColumn("_monotonically_increasing_id", monotonically_increasing_id())
+      .withColumn("_filename", input_file_name().as("_filename"))
+      .withColumn("_index", row_number().over(window).as("_index"))
+      .drop("_monotonically_increasing_id")
+
+    assert(TestUtils.datasetEquality(expected, actual))
+
   }
 
   test("TextExtract: Structured Streaming") {